requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
python-multipart
PyPDF2
python-docx
Pillow
PyMuPDF
